Remove OpenAPIConfig class, add API config annotations to main class
The OpenAPIConfig class, which was used to enforce that the API documentation server URLs use HTTPS, has been deleted. The functionality of defining server URLs has now been moved to UserApiApplication class by the use of @Server annotations. It now just sets the URL to "/" which magically makes Swagger-UI use the server root of its own URL as the URL for requests.